Move model handling to AbstractTemplate
